Test System Lambda with different Java versions
Newer JDK versions strongly encapsulate JDK internals. This includes the
map of environment variables. From JDK 16 on System Lambda's
withEnvironmentVariable does not work out of the box anymore. It
requires that the JVM is run with specific arguments. The new tests are
added so that I can verify that System Lambda still works with newer
JDKs.